+++ src/main/java/GUI/CompradorVents/MostrarServiciosVendedoresGUI.form
@@ -11,58 +11,62 @@
     </properties>
     <border type="none"/>
     <children>
-      <grid id="39753" binding="superiorPNL" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+      <grid id="39753" binding="superiorPNL" layout-manager="GridLayoutManager" row-count="3" column-count="8" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
         <margin top="0" left="0" bottom="0" right="0"/>
         <constraints>
           <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
         </constraints>
-        <properties/>
+        <properties>
+          <background color="-16757134"/>
+        </properties>
         <border type="none"/>
         <children>
-          <scrollpane id="5ab28">
+          <scrollpane id="5ab28" binding="vendedoresSCROLL">
             <constraints>
-              <grid row="1" column="0" row-span="3" col-span="1" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
+              <grid row="1" column="0" row-span="2" col-span="3" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties/>
             <border type="none"/>
             <children>
-              <component id="3e140" class="javax.swing.JList" binding="list1" default-binding="true">
+              <component id="3e140" class="javax.swing.JList" binding="vendedoresLIST">
                 <constraints/>
                 <properties>
                   <background color="-264257"/>
-                  <font name="Consolas" size="22" style="0"/>
+                  <font name="Consolas" size="18" style="0"/>
+                  <foreground color="-16316665"/>
                 </properties>
               </component>
             </children>
           </scrollpane>
           <component id="eac3" class="javax.swing.JLabel" binding="iconLBL">
             <constraints>
-              <grid row="0" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+              <grid row="0" column="0" row-span="1" col-span="8" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
               <icon value="icon-servicios.jpg"/>
               <text value=""/>
             </properties>
           </component>
-          <component id="70036" class="javax.swing.JComboBox" binding="comboBox1" default-binding="true">
+          <component id="70036" class="javax.swing.JComboBox" binding="filtroCBOX">
             <constraints>
-              <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
+              <grid row="1" column="3" row-span="1" col-span="4" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
+              <background color="-393274"/>
+              <font name="Consolas" size="20" style="0"/>
+              <foreground color="-16316665"/>
               <model/>
             </properties>
           </component>
-          <hspacer id="571f1">
+          <component id="95646" class="javax.swing.JButton" binding="filtroBTN">
             <constraints>
-              <grid row="3" column="1" row-span="1" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-            </constraints>
-          </hspacer>
-          <component id="95646" class="javax.swing.JButton" binding="button1" default-binding="true">
-            <constraints>
-              <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
+              <grid row="2" column="3" row-span="1" col-span="4" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
-              <text value="Button"/>
+              <background color="-4506567"/>
+              <font name="Consolas" size="20" style="0"/>
+              <foreground color="-16316665"/>
+              <text value="Filtrar"/>
             </properties>
           </component>
         </children>
@@ -72,37 +76,51 @@
         <constraints>
           <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
         </constraints>
-        <properties/>
+        <properties>
+          <background color="-12039237"/>
+        </properties>
         <border type="none"/>
         <children>
-          <component id="7d858" class="javax.swing.JComboBox" binding="comboBox2" default-binding="true">
+          <component id="7d858" class="javax.swing.JComboBox" binding="vendedorCBOX">
             <constraints>
               <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
-            <properties/>
+            <properties>
+              <background color="-6467331"/>
+              <font name="Consolas" size="20" style="0"/>
+              <foreground color="-16316665"/>
+            </properties>
           </component>
-          <component id="f1549" class="javax.swing.JButton" binding="button3" default-binding="true">
+          <component id="f1549" class="javax.swing.JButton" binding="buscarVendedorBTN">
             <constraints>
               <grid row="1" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
-              <text value="Button"/>
+              <background color="-4506567"/>
+              <font name="Consolas" size="20" style="0"/>
+              <foreground color="-264257"/>
+              <text value="Buscar Vendedor"/>
             </properties>
           </component>
-          <component id="b7e24" class="javax.swing.JButton" binding="button2" default-binding="true">
+          <component id="b7e24" class="javax.swing.JButton" binding="volverBTN">
             <constraints>
               <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
-              <text value="Button"/>
+              <background color="-4506567"/>
+              <font name="Consolas" size="20" style="0"/>
+              <foreground color="-264257"/>
+              <text value="Volver"/>
             </properties>
           </component>
-          <component id="ba926" class="javax.swing.JLabel">
+          <component id="ba926" class="javax.swing.JLabel" binding="vendeorLBL">
             <constraints>
               <grid row="0" column="1" row-span="1" col-span="2" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
-              <text value="Label"/>
+              <font name="Consolas" size="22" style="0"/>
+              <foreground color="-264257"/>
+              <text value="Seleccione el Número del vendedor que desea ver"/>
             </properties>
           </component>
         </children>
